Revamp AbstractCriterionBasedVariableSelector. - Replace TObjectDoubleMap by Weightable for efficiency reason - Dom/WDeg now manages futvars (via Element) for itself and Dom/WDegRef - remove AbstractCriterionBasedVariableSelector#increase (only required by Dom/WDeg and Dom/WDegRef)
